Move BaseConfigView to ilsgateway
